Better Multi-Level
Refactored so that we can now test possible nodes immediately instead of sending them to list.  This allows us to check if the node position up for testing was created from the same meshed object that the RayCasts hit to verify if the location is valid.  This allows multiple meshed objects to overlap one another and only a single node is generated instead of one for every object at that location.